implemented order table for responsive
